+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # [unreleased]
 
+# [v0.25.0] 2024-10-29
+
 ## Changed
 
 Renamed `PusFileSeqCountProvider` to `CcsdsFileSeqCountProvider` but keep old alias.